Remove HomeViewController and related code
