--- DuckDuckGo/HomePageDependencies.swift
@@ -1,37 +1,0 @@
-//
-//  HomePageDependencies.swift
-//  DuckDuckGo
-//
-//  Copyright © 2024 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-import Bookmarks
-import DDGSync
-import Core
-import BrowserServicesKit
-
-struct HomePageDependencies {
-    let homePageConfiguration: HomePageConfiguration
-    let model: Tab
-    let favoritesViewModel: FavoritesListInteracting
-    let appSettings: AppSettings
-    let syncService: DDGSyncing
-    let syncDataProviders: SyncDataProviders
-    let privacyProDataReporter: PrivacyProDataReporting
-    let variantManager: VariantManager
-    let newTabDialogFactory: any NewTabDaxDialogProvider
-    let newTabDialogTypeProvider: NewTabDialogSpecProvider
-}
