--- DuckDuckGo/hr.lproj/Home.strings
@@ -1,9 +1,0 @@
-/* Class = "UIView"; accessibilityLabel = "Search or enter address"; ObjectID = "7Jf-KE-oXj"; */
-"7Jf-KE-oXj.accessibilityLabel" = "Pretraži ili unesi adresu";
-
-/* Class = "UILabel"; text = "Search or enter address"; ObjectID = "73D-OT-n1K"; */
-"73D-OT-n1K.text" = "Pretraži ili unesi adresu";
-
-/* Class = "UIButton"; accessibilityLabel = "Settings"; ObjectID = "kcN-YB-mZi"; */
-"kcN-YB-mZi.accessibilityLabel" = "Postavke";
-
