--- DuckDuckGo/hu.lproj/Home.strings
@@ -1,9 +1,0 @@
-/* Class = "UIView"; accessibilityLabel = "Search or enter address"; ObjectID = "7Jf-KE-oXj"; */
-"7Jf-KE-oXj.accessibilityLabel" = "Keresd meg, vagy írd be a címet.";
-
-/* Class = "UILabel"; text = "Search or enter address"; ObjectID = "73D-OT-n1K"; */
-"73D-OT-n1K.text" = "Keresd meg, vagy írd be a címet.";
-
-/* Class = "UIButton"; accessibilityLabel = "Settings"; ObjectID = "kcN-YB-mZi"; */
-"kcN-YB-mZi.accessibilityLabel" = "Beállítások";
-
